Move dir for repository
