+++ README.md
@@ -162,10 +162,10 @@ $ go mod download
 
 - infra
   - sqlhandler: DBとのコネクションを確立する
+  - <model_name>: データストアからデータを取得するための処理を記述する。クエリ結果をモデルに(GoのStruct)にマッピングする。domain層のRepositoryで定義したinterfaceを満たすメソッドを持つ構造体を返す。
   - その他: マイグレーションスクリプトを配置する
 - domain:
   - model: ドメインモデルを定。データベースに保管されたカラムをプログラム上においてどのようなデータ構造で利用するかを記述する。モデルに紐付いたビジネスロジックも対象となる。
-- repository: データストアからデータを取得するための処理を記述する。クエリ結果をモデルに(GoのStruct)にマッピングし、モデルを返す
 - handler
   - router: リクエストとハンドラのマッピングをする(ルーティングの役割)
   - handler: ハンドラ関数(リクエストを受け取りレスポンスを返す関数)を記述する
