+++ api/infra/article.go
@@ -1,4 +1,4 @@
-package repository
+package infra
 
 import (
 	"database/sql"
